Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/webappapis/scripting/processing-model-2/compile-error-in-body-onerror-expected.txt

CONSOLE MESSAGE: SyntaxError: Unexpected token ')'
CONSOLE MESSAGE: SyntaxError: Unexpected end of script

PASS window.onerror - compile error in <body onerror>
